+++ NEWS.md
@@ -25,6 +25,8 @@
 - Fix new connection RStudio selectors colors when running
   under OS X Mojave.
 
+- Support for launching Livy logs from connection pane.
+
 - Removed `overwrite` parameter in `spark_read_table()` (#1698).
 
 - Fix regression preventing using R 3.2 (#1695).